Use Artifacts instead of Registry for moving CI Image
We would like to eventually move to a fork-based development model. However, PR's from fork's are unable to push images to the GHCR registry (they could push something bad). So instead, this changes CI to use artifacts instead to pass around the container image.

Surprisingly, this actually slightly reduces the time it takes to load & start the Docker container.