CI: Conditionally build the CI image.
If the Dockerfile's have been changed, we need to rebuild the
CI image. Else, we can skip this step and just re-use the
latest image in-sync with main.

Currently there is a workaround regarding how we pass the new
CI image name should it need to be built. This is a limitation
with how GitHub handles matrix job's with outputs. Because our
image tag is based on a constant, the jobs within the matrix
will all provide the same value at least. Ideally we would
provide the full image name and tag. We can look at refactoring
the workflow to avoid this situation.

If a new image needs to be built, it will also attempt to push this
image to the docker registry. This means any changes to the Dockerfile
will require a PR to be opened directly in hipFile instead of a fork.

Given that the Dockerfile is expected to rarely change, this should be
an acceptable trade-off.