Add caching to PR-image only.
This cache will only be used by the same PR if it gets updated
for whatever reason. This way people won't be sad if a particular
distro takes 20 minutes to download packages for if they need to
change the PR.

Intentionally not caching from 'latest' to have a relatively
up-to-date image. Open to discuss about this.