CI: Re-add caching for updating the 'latest' image.
On pull_requests, if the Dockerfile has been modified, a new image will be
build and pushed using as many cached layers as possible.
Otherwise, this effectively is a no-op as the entire image is already
cached.

If this is triggered manually via workflow_dispatch, this gives the option
of not consuming the cache to forcibly update the image packages.
Regardless of this option, the resulting image is cached for future use.