CI: Update container names
Note: Replaces '+' with 'plus' to make it safe for Docker to use.
Docker does not like most non-alphanumeric characters when it comes
to naming containers.

On a self-hosted machine with multiple CI runners, this could
cause matrix jobs to create a container with the same name.
(Matrix jobs do not modify job_id.)
This change was introduced when we moved to GitHub provided runners.
So while we have not hit this issue, we should still fix it.

Also makes some cosmetic changes to come container names like dropping
a redundant "_nvidia" suffix.