Move NVIDIA API compatibility tests
Move the hipfile-cufile.cpp file to a new nvidia_detail directory
and rename it to cufile-api-compat.cpp to make its purpose
more obvious.

Also added some explanatory text to the test file.

Also removed the useless main.cpp file that was a leftover
placeholder from when we first set up the testing.