Use default dtors for BufferMap and FileMap
Making virtual calls in destructors is a bad idea in C++ and
clang-tidy complains about this. There's really no reason to
walk the maps to free elements, as the elements are all shared
pointers and will be cleaned up automatically as the reference
counts drop to zero.

This PR replaces the explicit destructors with default
destructors.