Remove virtual calls in dtors
clang-tidy complains about using virtual calls in the FileMap and
BufferMap destructors. The virtual calls in question attempt to
shut down the map objects, which is unnecessary, as the maps
contain shared pointers that will be cleaned when they go out
of scope.

The functionality that we retained, was complaining about any in-use
File or Buffer objects when the associated maps are shut down.
Previously, this would result in exit() being called in debug builds
and silently ignoring the problem in release builds. The new behavior
is to create syslog messages in all builds (like our Stream class).