CI: Capture package metadata to use elsewhere in CI.
We would like to capture package metadata to use in other
places in CI. For example, specifying the correct package to
download in later steps. CI is unable to poll this data before
compilation, so we cannot construct the exact filename
beforehand. Besides, different platform's will produce
a different naming convention for the packages. We assume only a
single package is produced, so using a glob pattern should be safe
here.

For now we specifically care about the version. The gate listed
has been tested and if an empty string is returned, the
validation step will fail. This has been tested by introducing
a temp commit to force an empty string to be returned.