CI: Mount the GitHub Actions CMD files into the container.
These files are used to allow a job/step to pass information
to other jobs/steps. They are usually specified through
the ${GITHUB_OUTPUT} and ${GITHUB_ENV} environment variables
(note there are others), which reference a path to a file
on the host runner. It is important to note: these files
are unique on each step of a job. That is, the file
pointed to changes on each step.

To work with this, we mount the entire directory that contains
these files into the container, and then when needed, parse
the filename out of ${GITHUB_*} to reference it from within
the container. This allows us to more finely control adding
information to these files rather than catching the TTY output
from the entire Docker container.