CI: Use cancelled() instead of always()
While rather poorly documented, using always() in a condition
of a job that calls another workflow causes that workflow to not
be cancellable. You will not be able to cancel that job normally
or by updating the PR. Only a force-cancel will work.

Using !cancelled() instead provides the same logic, but removes
the special behaviour that always() provides.