Turn off sanitizers in executables
This is a bit crude, but GTest has severe problems when the test
applications are built with the sanitizers.

This change will disable compiling the test applications with the
sanitizer flags. It will still compile the library with sanitizers
and we still need to link the test programs with the sanitizer
libraries.