CI: Capture and save the produced package filename.
CI does not have a way to tell what the resulting filename of
the built package is going to be prior to the package being
built. This is because we do not have a way to get access to
what the build version, release, or target arch is. If we
had this, we could construct the expected filename based on the
DEB/RPM convention.

Instead, we assume that only a single DEB/RPM will be built and
that we can just glob for the filename. This change gets the
full filename once, and then stores it as an output so we do
not need to continue to glob for the filename of the package.